refactor: icon-map.tsx를 icon-entries.ts로 통합(#5)
